+++ scratch.js
@@ -9,8 +9,10 @@ const readFile = require('./tests/lib/_cachedPage');
 // var str = `{{ill|Joke|fr|Blague|hu|Vicc|de|Witz}}`;
 // var str = `hello {{small|2 February}}`;
 // var str = `{{tiw|Hatnote}}`;
-var str = `{{date|June 8 2018|mdy}}`;
-var str = `{{l|cs|háček}}`;
+// var str = `{{date|June 8 2018|mdy}}`;
+// var str = `{{l|cs|háček}}`;
+var str = `{{Coord|44.112|-87.913|display=title}}`;
 let doc = wtf(str);
 console.log(doc.plaintext());
 console.log(doc.templates());
+console.log(doc.coordinates());
