+++ src/templates/geo/index.js
