+++ src/templates/inline/index.js
