+++ src/templates/parsers/inside.js
