+++ src/templates/parsers/pipeSplit.js
